anve-customer-service-ai/requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-multipart>=0.0.5,<0.1.0
sqlalchemy>=1.4.0,<2.0.0
psycopg2-binary>=2.9.1,<3.0.0
pydantic>=1.8.0,<2.0.0
langchain>=0.0.200
openai>=0.27.0
python-dotenv>=0.19.0
streamlit>=1.22.0
pytest>=7.0.0
httpx>=0.24.0
python-jwt>=4.0.0
rate-limiter>=3.0.0
tenacity>=8.2.2
loguru>=0.7.0 